api/requirements.txt
azure-functions
azure-identity
azure-keyvault-secrets
azure-storage-blob
firebase-admin
joblib
pandas
numpy
scikit-learn